Move adding strings to WebUI out of chromium_src.
Fixes brave/brave-browser#22944
